Remove de-german/build directory
We are no longer keeping the build files but rather expecting the CSVs
to be edited and the build files to be generated from them. When working
the other way around that is fine but the builds would be considered
temporary.

No data is lost because it can be re-generated from the de-german/src/
directory.