Elaborate on how to make audio recordings
This is a suggestion for issue #6. Take the text from the discussion there and some
additional information and edit it a bit.

In an ideal world licensing information should be listed here but it is not
clear what exactly (perhaps CC-BY or similar?). Perhaps also add one of those
websites that explain the concept of licensing and the different options
available, in case the reader is not familar with open-source. Put ??? here for
the time being.