Clarify who does the audio post-processing
Contributors who do not know how to do these steps aftewards might be a bit
intimidated by these comments. Change the phrasing so that the it is explicit
that people who wrote the README will take care of this. People who are able
and willing to do this themselves will probably first get in contact with us if
it is phrased like this too.