docs: improve the contribution guidelines
* Restructure the text and add headings where applicable. Use semantic line
breaks for slightly nicer diffs (See https://sembr.org/).

* Turn e-mail addresses into links to save the reader from copying the text
into their e-mail clients.

* Suggest CC BY-SA 4.0 as a licence for audio contributions. Also make
contributing less intimidating for non-programmer not familar with open-source
by linking link to an informational website about software licencing in general
which explains the concept of licensing and the different options available

* Elaborate on how to make audio recordings with the text from the discussion
under issue #6.

* Clarify who does the audio post-processing by changing the phrasing so that
the it is explicit that people who wrote the README will take care of the audio
post-processing. Contributors who are able and willing to do this themselves
will probably first get in contact with us if it is phrased like this too.