tests: failing file name normalisation
These are preparations for issue #18. Maybe this code is slightly overkill and
a bit too repetitive, but in general re-importing makes for more robust tests.
Testing private functions is also bad form, but the public interface
(`generate_audio`) relies on the file system and Google API calls. This code
could be improved with `pytest`'s temporary directory fixtures, and perhaps
`google.cloud.texttospeech` could be monkey-patched as well. I am not sure how
this labour compares to a refactor, but the latter might be easier with more
tests like these.